requirements.txt
Django==5.2.7
djangorestframework==3.16.1
django-cors-headers==4.9.0
pandas==2.3.3
numpy==2.3.4
openpyxl==3.1.5
python-decouple==3.8
